Separated executable server from server model
