+++ inference/__init__.py
@@ -8,17 +8,17 @@ from .yolo_world import YOLOWorld
 from .detection_result import Detection, BoundingBox, FrameDetections
 from .results_writer import ResultsWriter
 from .visualization import FrameAnnotator, AnnotatedVideoWriter
-from .server import YOLOWorldServer, host_model, str_to_image
+from .server_model import YOLOWorldServer
 
 __all__ = [
     'YOLOWorld',
+    'YOLOWorldServer',
     'Detection',
     'BoundingBox',
     'FrameDetections',
     'ResultsWriter',
     'FrameAnnotator',
     'AnnotatedVideoWriter',
-    'YOLOWorldServer',
     'host_model',
     'str_to_image'
 ]
